requirements.txt
tensorflow==2.10.0
keras==2.10.0
opencv-Python==4.11.0.86
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.6.1
imageio==2.37.0
einops==0.8.1
tqdm==4.67.1
ffmpeg-python==0.2.0
tensorflow-hub==0.16.1
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-genai==1.9.0
google-pasta==0.2.0